Rename `store` to `stores`
