+++ src/components/admin/AdminToolBar.vue
@@ -50,8 +50,8 @@
 
 <script lang="ts" setup>
 import { logout } from "@/api/login";
-import { appMenu } from "@/store/app-store";
-import { useUserStore } from "@/store/user";
+import { appMenu } from "@/stores/app-store";
+import { useUserStore } from "@/stores/user";
 
 const $router = useRouter();
 const $route = useRoute();
