+++ src/views/login/Login.vue
@@ -56,7 +56,7 @@
   </div>
 </template>
 <script setup lang="ts">
-import { useUserStore } from "@/store/user";
+import { useUserStore } from "@/stores/user";
 import Message from "@/utils/message";
 
 const $router = useRouter();
